+++ CHANGELOG.md
@@ -9,7 +9,7 @@ You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
 ---------------------------------------------------------------------------
-## 1.2.6 - 2022-10-??
+## 1.2.6 - 2022-10-08
 
 Incremental improvements and bug fix.
 
@@ -18,6 +18,8 @@ Incremental improvements and bug fix.
 
 ### Bug fixes:
 - Fix issue with S3 to LINUX transfer. Destination path was being constructed incorrectly.
+- Fix issue with S3 path contained in FileInfo. Should be relative path and not absolute path.
+- Fix issues with populating path and url in FileInfo objects for all system types.
 
 ---------------------------------------------------------------------------
 ## 1.2.5 - 2022-09-26